Narrow, chunk and split are all high-level slice-based methods
