+++ crates/burn-autodiff/src/ops/int_tensor.rs
@@ -265,39 +265,6 @@ impl<B: Backend, C: CheckpointStrategy> IntTensorOps<Self> for Autodiff<B, C> {
         B::int_swap_dims(tensor, dim1, dim2)
     }
 
-    fn int_narrow(
-        tensor: <Autodiff<B> as Backend>::IntTensorPrimitive,
-        dim: usize,
-        start: usize,
-        length: usize,
-    ) -> <Autodiff<B> as Backend>::IntTensorPrimitive {
-        B::int_narrow(tensor, dim, start, length)
-    }
-
-    fn int_chunk(
-        tensor: <Autodiff<B> as Backend>::IntTensorPrimitive,
-        chunks: usize,
-        dim: usize,
-    ) -> Vec<<Autodiff<B> as Backend>::IntTensorPrimitive> {
-        B::int_chunk(tensor, chunks, dim)
-    }
-
-    fn int_split(
-        tensor: <Autodiff<B> as Backend>::IntTensorPrimitive,
-        split_size: usize,
-        dim: usize,
-    ) -> Vec<<Autodiff<B> as Backend>::IntTensorPrimitive> {
-        B::int_split(tensor, split_size, dim)
-    }
-
-    fn int_split_with_sizes(
-        tensor: <Autodiff<B> as Backend>::IntTensorPrimitive,
-        split_sizes: Vec<usize>,
-        dim: usize,
-    ) -> Vec<<Autodiff<B> as Backend>::IntTensorPrimitive> {
-        B::int_split_with_sizes(tensor, split_sizes, dim)
-    }
-
     fn int_random(
         shape: Shape,
         distribution: Distribution,
