+++ crates/burn-tensor/src/tensor/api/mod.rs
@@ -5,26 +5,20 @@ mod autodiff;
 mod base;
 mod bool;
 mod cartesian_grid;
-mod chunk;
 mod float;
 mod int;
 mod kind;
-mod narrow;
 mod numeric;
 mod slice;
 mod sort;
-mod split;
 mod transaction;
 
 pub use argwhere::argwhere_data;
 pub use autodiff::*;
 pub use base::*;
 pub use cartesian_grid::cartesian_grid;
-pub use chunk::chunk;
 pub use kind::*;
-pub use narrow::narrow;
 pub use numeric::*;
 pub use slice::*;
 pub use sort::{argsort, sort, sort_with_indices};
-pub use split::{split, split_with_sizes};
 pub use transaction::*;
