+++ crates/burn-tensor/src/tensor/ops/bool_tensor.rs
@@ -3,10 +3,10 @@ use super::{
     repeat_dim::repeat_with_slice_assign,
 };
 use crate::{
-    Bool, ElementConversion, TensorData, TensorMetadata, argwhere_data, backend::Backend, chunk,
-    narrow, split, split_with_sizes, tensor::Shape,
+    Bool, ElementConversion, TensorData, TensorMetadata, argwhere_data, backend::Backend,
+    tensor::Shape,
 };
-use alloc::{vec, vec::Vec};
+use alloc::vec::Vec;
 use core::{future::Future, ops::Range};
 
 /// Bool Tensor API for basic operations, see [tensor](crate::Tensor)
@@ -263,80 +263,6 @@ pub trait BoolTensorOps<B: Backend> {
     /// The tensor with the elements reversed.
     fn bool_flip(tensor: BoolTensor<B>, axes: &[usize]) -> BoolTensor<B>;
 
-    /// Returns a new tensor with the given dimension narrowed to the given range.
-    ///
-    /// # Arguments
-    ///
-    /// * `dim` - The dimension along which the tensor will be narrowed.
-    /// * `start` - The starting point of the given range.
-    /// * `length` - The ending point of the given range.
-    /// # Panics
-    ///
-    /// - If the dimension is greater than the number of dimensions of the tensor.
-    /// - If the given range exceeds the number of elements on the given dimension.
-    ///
-    /// # Returns
-    ///
-    /// A new tensor with the given dimension narrowed to the given range.
-    fn bool_narrow(
-        tensor: BoolTensor<B>,
-        dim: usize,
-        start: usize,
-        length: usize,
-    ) -> BoolTensor<B> {
-        narrow::<B, Bool>(tensor, dim, start, length)
-    }
-
-    /// Split the tensor along the given dimension into chunks.
-    ///
-    /// # Arguments
-    ///
-    /// * `tensor` - The tensor.
-    /// * `chunks` - The number of chunks to be produced.
-    /// * `times` - The dimension along which the tensor will be split.
-    ///
-    /// # Returns
-    ///
-    /// A vector of tensors.
-    fn bool_chunk(tensor: BoolTensor<B>, chunks: usize, dim: usize) -> Vec<BoolTensor<B>> {
-        chunk::<B, Bool>(tensor, chunks, dim)
-    }
-
-    /// Split the tensor along the given dimension into chunks of `split_size`.
-    ///
-    /// # Arguments
-    ///
-    /// * `tensor` - The tensor.
-    /// * `split_size` - The size of a single chunk.
-    /// * `times` - The dimension along which the tensor will be split.
-    ///
-    /// # Returns
-    ///
-    /// A vector of tensors.
-    fn bool_split(tensor: BoolTensor<B>, split_size: usize, dim: usize) -> Vec<BoolTensor<B>> {
-        split::<B, Bool>(tensor, split_size, dim)
-    }
-
-    /// Split the tensor along the given dimension into chunks with sizes in
-    /// `dim` according to `split_sizes`.
-    ///
-    /// # Arguments
-    ///
-    /// * `tensor` - The tensor.
-    /// * `split_sizes` - Vector of sizes for each chunk.
-    /// * `times` - The dimension along which the tensor will be split.
-    ///
-    /// # Returns
-    ///
-    /// A vector of tensors.
-    fn bool_split_with_sizes(
-        tensor: BoolTensor<B>,
-        split_sizes: Vec<usize>,
-        dim: usize,
-    ) -> Vec<BoolTensor<B>> {
-        split_with_sizes::<B, Bool>(tensor, split_sizes, dim)
-    }
-
     /// Tests if any element in the boolean `tensor` evaluates to True.
     ///
     /// # Arguments
@@ -419,35 +345,6 @@ pub trait BoolTensorOps<B: Backend> {
             let device = B::bool_device(&tensor);
             let data = B::bool_into_data(tensor).await;
             argwhere_data::<B>(data, &device)
-        }
-    }
-
-    /// Compute the indices of the elements that are non-zero.
-    ///
-    /// # Arguments
-    ///
-    /// * `tensor` - The input tensor.
-    ///
-    /// # Returns
-    ///
-    /// A vector of tensors, one for each dimension of the given tensor, containing the indices of
-    /// the non-zero elements in that dimension. If all elements are zero, the vector is empty.
-    fn bool_nonzero(
-        tensor: BoolTensor<B>,
-    ) -> impl Future<Output = Vec<IntTensor<B>>> + 'static + Send {
-        async {
-            let indices = B::bool_argwhere(tensor).await;
-
-            if indices.shape().num_elements() == 0 {
-                // Return empty vec when all elements are zero
-                return vec![];
-            }
-
-            let dims = indices.shape().dims;
-            B::int_chunk(indices, dims[1], 1)
-                .into_iter()
-                .map(|t| B::int_reshape(t, Shape::new([dims[0]])))
-                .collect()
         }
     }
